Put all QoS coding into one file
