--- felix/dataplane/linux/bwp/bandwidth.go
@@ -1,54 +1,0 @@
-// Copyright 2018 CNI authors
-// Changes copyright (c) 2025 Tigera, Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package bwp
-
-import (
-	"github.com/containernetworking/plugins/pkg/utils"
-	"github.com/vishvananda/netlink"
-)
-
-const maxIfbDeviceLength = 15
-const ifbDevicePrefix = "bwcali"
-
-func SafeQdiscList(link netlink.Link) ([]netlink.Qdisc, error) {
-	qdiscs, err := netlink.QdiscList(link)
-	if err != nil {
-		return nil, err
-	}
-	result := []netlink.Qdisc{}
-	for _, qdisc := range qdiscs {
-		// filter out pfifo_fast qdiscs because
-		// older kernels don't return them
-		_, pfifo := qdisc.(*netlink.PfifoFast)
-		if !pfifo {
-			result = append(result, qdisc)
-		}
-	}
-	return result, nil
-}
-
-func GetMTU(deviceName string) (int, error) {
-	link, err := netlink.LinkByName(deviceName)
-	if err != nil {
-		return -1, err
-	}
-
-	return link.Attrs().MTU, nil
-}
-
-func GetIfbDeviceName(deviceName string) string {
-	return utils.MustFormatHashWithPrefix(maxIfbDeviceLength, ifbDevicePrefix, deviceName)
-}
